+++ synchronizer/src/sync_loop.rs
@@ -8,8 +8,6 @@ use tokio::sync::watch;
 use tracing::{debug, info, warn};
 
 use crate::node::{Node, ProcessedSlot};
-use crate::sync_db::CommittedSlotRecord;
-
 const HEAD_CHECK_INTERVAL: Duration = Duration::from_secs(12);
 
 enum SlotHeaderState {
@@ -228,11 +226,23 @@ pub(crate) async fn initialize_sync(
     info!(head_slot = head.slot, head_root = ?head.root, "Fetched initial beacon head");
 
     let bootstrap_start_slot = initial_start_slot.unwrap_or(head.slot);
-    let bootstrap_slot = load_bootstrap_slot_record(node, &head, bootstrap_start_slot).await?;
+    let bootstrap_slot = bootstrap_start_slot.checked_sub(1).ok_or_else(|| {
+        anyhow!("bootstrap start slot must be > 0 to insert initial canonical row")
+    })?;
+
+    if bootstrap_slot > head.slot {
+        return Err(anyhow!(
+            "INITIAL_START_SLOT {bootstrap_start_slot} is ahead of current beacon head {}; cannot bootstrap slot {}",
+            head.slot,
+            bootstrap_slot
+        ));
+    }
+
+    let bootstrap_record = node.load_committed_slot_record(bootstrap_slot).await?;
 
     let start_slot = node
         .sync_db
-        .ensure_bootstrap_row(bootstrap_slot)
+        .ensure_bootstrap_row(bootstrap_record)
         .await?
         .checked_add(1)
         .ok_or_else(|| anyhow!("last processed slot overflow"))?;
@@ -246,58 +256,6 @@ pub(crate) async fn initialize_sync(
     Ok(SyncStart {
         next_slot: start_slot,
         head,
-    })
-}
-
-async fn load_bootstrap_slot_record(
-    node: &Node,
-    current_head: &BlockHeader,
-    start_slot: u32,
-) -> Result<CommittedSlotRecord> {
-    let bootstrap_slot = start_slot.checked_sub(1).ok_or_else(|| {
-        anyhow!("bootstrap start slot must be > 0 to insert initial canonical row")
-    })?;
-
-    if bootstrap_slot > current_head.slot {
-        return Err(anyhow!(
-            "INITIAL_START_SLOT {start_slot} is ahead of current beacon head {}; cannot bootstrap slot {}",
-            current_head.slot,
-            bootstrap_slot
-        ));
-    }
-
-    let Some(header) = node
-        .get_beacon_slot_header_with_retry(bootstrap_slot)
-        .await?
-    else {
-        return Ok(CommittedSlotRecord {
-            slot: bootstrap_slot,
-            block_root: None,
-            parent_root: None,
-            block_number: None,
-            current_gsr: None,
-            is_empty: true,
-        });
-    };
-
-    let block = node
-        .get_beacon_block_by_hash_with_retry(bootstrap_slot, header.root)
-        .await?;
-
-    let execution_payload = block.execution_payload.as_ref().ok_or_else(|| {
-        anyhow!(
-            "Beacon block {} for bootstrap slot {bootstrap_slot} had no execution payload",
-            header.root
-        )
-    })?;
-
-    Ok(CommittedSlotRecord {
-        slot: bootstrap_slot,
-        block_root: Some(header.root),
-        parent_root: Some(block.parent_root),
-        block_number: Some(execution_payload.block_number),
-        current_gsr: None,
-        is_empty: false,
     })
 }
